refactor: remove unused code
