--- src/routes/priceRoutes.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import { getNegativePricePeriods } from "../controllers/priceController";
-
-const router = Router();
-
-router.get("/negative", getNegativePricePeriods);
-
-export default router;
