--- src/services/consumptionService.ts
@@ -1,8 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-
-const prisma = new PrismaClient();
-
-export const getSomething = async () => {
-  // Implement the logic to
-  return [];
-};
