--- src/services/priceService.ts
@@ -1,8 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-
-const prisma = new PrismaClient();
-
-export const getNegativePricePeriodsService = async () => {
-  // Implement the logic to get the longest negative price periods
-  return [];
-};
